+++ db/edit.php
@@ -141,3 +141,4 @@ if ('save' === $_action) {
 </div>
 
 <?php
+